Simple string without args opti and benchmarker finished

+++ include/Flags.h
@@ -6,7 +6,7 @@
 /*   By: cberganz <cberganz@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/23 21:42:08 by cberganz          #+#    #+#             */
-/*   Updated: 2022/10/27 02:53:18 by charles          ###   ########.fr       */
+/*   Updated: 2022/10/29 05:42:40 by charles          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@
 # include <ctype.h>
 
 typedef struct s_flags {
-	uint16_t	flags;
+	uint8_t		flags;
 	uint32_t	field_width;
 	uint32_t	precision;
